refactor(theatre): relocate screencast util to src/ (core layer, used by crawler)
